Switch to sending Facebook Access Tokens to the API for Authentication (#18)
This pull request changes the passport strategy to facebookTokenStrategy instead of facebookStrategy. With facebookTokenStrategy the API accepts a facebook access token from the client. This lets us decouple the frontend from the backend so that the API is truly independent. Before this, the frontend would have to share the same state as the backend, or the API would have to have a user authenticate by going through the Facebook permission screen. Now, the user can authenticate on the frontend only, and the frontend can send the access token to the authentication endpoint (see test plan for details).