Config for Heroku environments (#21)
All of the changes in this PR are to get the code working in all heroku environments!

I still need to do a little research on webpack and webpack dev server and figure out how to make that work better, but for now, we can just run the webpack command (literally just type in webpack in the HAH directory) to build the client locally after changes are made. It's a little more of a pain than having wds restart automatically when changes are made, but this will help us to get verifiers started and then we can make local improvements later on.