Add PrimaryContactPerson, closedAllDay, bug fixes in front-end/back-end testing, new data for testing (#37)
changes PrimaryContact to PrimaryContactPerson in all cases
adds in closedAllDay as part of hours (open-now endpoint will need to change in the future to accommodate this)
updates the data import process to match new schemas (yarn data)
rewrites database-helpers.js to fix issues that were preventing the front-end from saving primaryContactPerson and verifiedData correctly, with added bonus of better readability
adds a new command for front-end/back-end testing: yarn test-data. yarn test-data drops everything from the database and imports one pregnancy center that has nearly all the possible data filled in with fake data, so that it's easy to see what isn't displaying on the front-end, and then also, what isn't saving on the backend.