Clear form on save (#44)
After a resource has been saved, we need to clear out the form so that when a new resource is loaded, none of the old data is included. This adds a button to clear the values. Note: this is not the most ideal way to clear the form, but with our upcoming enhancements, we will change this behavior.