adds inVerification field (#30)
Adds inVerification field to the pregnancy center schemas, updates the field to the userId when /verify is called. Includes a script to clear the inVerification field as needed. Tests updated to cover the new functionality. On logout, any pregnancy centers with a inVerification value that corresponds to that user can be released for more verification. Added a Joi ObjectId validator.