Remove primary contact fields for now (#66)
In order to keep things simple for the volunteers who are verifying data, we are not going to have them ask for a primary contact right now. Therefore, I commented out the primary contact fields on the form. I decided to comment them out, rather than remove them, because eventually we will want to add it back. We also don't need to remove any code on the server side, as everything still works, the primary contact information is just ignored.