Routing for mini-app (#98)
The first scaffolding PR for the Mini App!!

React scaffolding for mini app
Add React routing
Add data to redux
Display data on page
Note: there are some things that are hard coded for now. All of these files will eventually be updated with styling and to match the designs!